smoke-alarm-repo/agent/requirements.txt
typer==0.12.3
python-dotenv==1.0.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.142.0
pandas==2.2.2
requests==2.32.3
protobuf==4.25.3
orjson==3.10.7
geopy==2.4.1
